tmf: Delete linked trace elements when workspace trace resource deleted
Automatically delete trace element that is a link to a workspace
location when the resource at this workspace location is deleted.

This does not delete the trace element when it is a link to a location
outside the workspace that is deleted.

Reviewed-on: https://git.eclipse.org/r/110169
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>

+++ tmf/org.eclipse.tracecompass.tmf.ui/src/org/eclipse/tracecompass/tmf/ui/project/model/TmfExperimentElement.java
@@ -431,7 +431,7 @@ public class TmfExperimentElement extends TmfCommonProjectElement implements IPr
         resource.delete(true, null);
         IContainer parent = resource.getParent();
         // delete empty folders up to the parent experiment folder
-        if (!parent.equals(getResource()) && parent.members().length == 0) {
+        if (!parent.equals(getResource()) && parent.exists() && parent.members().length == 0) {
             deleteTraceResource(parent);
         }
     }
